Move more of orig-scope mapping into utils/
