+++ src/utils/pause/mapScopes/findGeneratedBindingFromPosition.js
@@ -21,7 +21,7 @@ import type {
   BindingContents
 } from "../../../types";
 // eslint-disable-next-line max-len
-import type { GeneratedBindingLocation } from "../../../actions/pause/mapScopes";
+import type { GeneratedBindingLocation } from "./index";
 
 import { createObjectClient } from "../../../client/firefox";
